Add set name method to the path interface and concern and remove it from main class
